[ALCYN-1494] Don't try to delete unexisting lines

+++ odoo/local-src/db2_import/converter/purchase.py
@@ -235,7 +235,8 @@ class DB2MapperPurchaseOrder(object):
                 row['ecfsui'], int(row['ecffou']),
                 int(row['ecfsuc']), int(line['dcfnli']))
             line_rec = rec.env.ref(xmlid, raise_if_not_found=False)
-            line_rec.unlink()
+            if line_rec:
+                line_rec.unlink()
 
         is_received = all(received_lines)
         is_done = is_received or expired
